Pbadillo

+++ SmartBoard.Data/Models/SmartBoard/TblObraconcepto.cs
@@ -11,7 +11,7 @@ namespace SmartBoard.Data.Models.SmartBoard
         public string Concepto { get; set; }
         public string Unidad { get; set; }
         public decimal PrecioUnitario { get; set; }
-        public int Cantidad { get; set; }
+        public decimal Cantidad { get; set; }
         public decimal Importe { get; set; }
         public bool Activo { get; set; }
         public DateTime Fecha { get; set; }
